Prevent navigation guard from misidentifying user auth state on app load
Adding waitUntilReady() action in the authStore to use it in navigation guards to observe the readiness of the store before reading its state. this resloves the bug of always considering the user as unauthenticated while the app load which results into redirecting him to /login if he tries to navigate to a protected route directly via the browser e.g. "/auctions/create".